+++ packtools/stylechecker.py
@@ -123,22 +123,25 @@ def _main():
     parser = argparse.ArgumentParser(
             description='SciELO PS stylechecker command line utility.',
             epilog=EPILOG)
-    parser.add_argument('--annotated', action='store_true',
-                        help='reproduces the XML with notes at elements that have errors')
+
+    mutex_group = parser.add_mutually_exclusive_group()
+    mutex_group.add_argument('--annotated', action='store_true',
+                             help='reproduces the XML with notes at elements that have errors')
+    mutex_group.add_argument('--raw', action='store_true',
+                             help='each result is encoded as json, without any formatting, and written to stdout in a single line.')
+
     parser.add_argument('--nonetwork', action='store_true',
                         help='prevents the retrieval of the DTD through the network')
     parser.add_argument('--assetsdir', default=None,
                         help='lookup, at the given directory, for each asset referenced by the XML. current working directory will be used by default.')
-    parser.add_argument('XML', nargs='*',
-                        help='filesystem path or URL to the XML')
     parser.add_argument('--version', action='version', version=packtools_version)
-    parser.add_argument('--loglevel', default='WARNING')
+    parser.add_argument('--loglevel', default='ERROR')
     parser.add_argument('--nocolors', action='store_false',
                         help='prevents the output from being colorized by ANSI escape sequences')
-    parser.add_argument('--noaggregation', action='store_true',
-                        help='prevents the output from being aggregated as a single json data structure')
     parser.add_argument('--extrasch', default=None,
                         help='runs an extra validation using an external schematron schema.')
+    parser.add_argument('XML', nargs='*',
+                        help='filesystem path or URL to the XML')
     args = parser.parse_args()
 
     logging.basicConfig(level=getattr(logging, args.loglevel.upper()))
@@ -192,7 +195,7 @@ def _main():
                     xml_validator.validate()[0] and
                     xml_validator.validate_style()[0])
 
-            if args.noaggregation:
+            if args.raw:
                 print(json.dumps(summary, sort_keys=True))
             else:
                 summary_list.append(summary)